override default scaffold generator in rails
